Blogger/requirements.txt
google-adk>==1.0.0